+++ tests/Validator/Type/IntegerValidatorTest.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Validator\Type;
@@ -10,16 +11,33 @@ class IntegerValidatorTest extends TestCase
     /**
      * Valid.
      *
-     * Test that integer value '9' is a valid integer.
+     * Test that integer value '-9' is a valid integer.
      *
+     * @covers \ExtendsSoftware\ExaPHP\Validator\Type\IntegerValidator::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Validator\Type\IntegerValidator::validate()
      */
     public function testValid(): void
     {
         $validator = new IntegerValidator();
-        $result = $validator->validate(9);
+        $result = $validator->validate(-9);
 
         $this->assertTrue($result->isValid());
+    }
+
+    /**
+     * Unsigned.
+     *
+     * Test that integer value '-9' is not a valid unsigned integer.
+     *
+     * @covers \ExtendsSoftware\ExaPHP\Validator\Type\IntegerValidator::__construct()
+     * @covers \ExtendsSoftware\ExaPHP\Validator\Type\IntegerValidator::validate()
+     */
+    public function testUnsigned(): void
+    {
+        $validator = new IntegerValidator(true);
+        $result = $validator->validate(-9);
+
+        $this->assertFalse($result->isValid());
     }
 
     /**
@@ -27,6 +45,7 @@ class IntegerValidatorTest extends TestCase
      *
      * Test that string value 'foo' is an valid integer.
      *
+     * @covers \ExtendsSoftware\ExaPHP\Validator\Type\IntegerValidator::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Validator\Type\IntegerValidator::validate()
      * @covers \ExtendsSoftware\ExaPHP\Validator\Type\IntegerValidator::getTemplates()
      */