+++ CHANGELOG.md
@@ -54,6 +54,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Changed BetweenValidator internal value validator from numeric to integer.
 - Changed the order of commands alphabetically when described by the shell descriptor.
 - Changed that a POST request body is only JSON decoded when the Content-Type is set with the value application/json.
+- Changed IntegerValidator to accept unsigned flag to allow only unsigned values.
 
 ### Deprecated